+++ lib/see_more.dart
@@ -1,3 +1,3 @@
 export 'src/see_more_controller.dart' show SeeMoreController;
-export 'src/see_more_widget.dart' show SeeMoreWidget;
+export 'src/see_more_widget.dart' show SeeMoreWidget, SeeMoreAnnotation;
 export 'src/trim_mode.dart' show TrimMode;